Create Delete Node in a Binary Search Tree (BST)
Given the root node of a binary search tree and a key, remove the key from the tree and return the root node.

Note: The tree contains only unique values. Do not assume that the key is present in the BST.

There can be multiple solutions. You can return the root node containing any of the solutions as long as the tree is a BST.